System.Random

Description

This library deals with the common task of pseudo-random number generation.

Overview

This library provides type classes and instances for the following concepts:

Pure pseudo-random number generators

RandomGen is an interface to pure pseudo-random number generators.
StdGen, the standard pseudo-random number generator provided in this library, is an instance of RandomGen. It uses the SplitMix implementation provided by the splitmix package. Programmers may, of course, supply their own instances of RandomGen.

Monadic pseudo-random number generators

MonadRandom is an interface to monadic pseudo-random number generators.

Monadic adapters

Frozen, PrimGen and MutGen turn a RandomGen instance into a MonadRandom instance.

Drawing from a range

UniformRange is used to generate a value of a datatype uniformly within a range.
This library provides instances of UniformRange for many common numeric datatypes.

Drawing from the entire domain of a type

Uniform is used to generate a value of a datatype uniformly over all possible values of that datatype.
This library provides instances of Uniform for many common bounded numeric datatypes.

Backwards compatibility and deprecations

Version 1.2 mostly maintains backwards compatibility with version 1.1. This has a few consequences users should be aware of:

The type class Random is deprecated and only provided for backwards compatibility. New code should use Uniform and UniformRange instead.

The methods next and genRange in RandomGen are deprecated and only provided for backwards compatibility. New instances of RandomGen should implement word-based methods instead. See below for more information about how to write a RandomGen instance.

This library provides instances for Random for some unbounded datatypes for backwards compatibility. For an unbounded datatype, there is no way to generate a value with uniform probability out of its entire domain, so the random implementation for unbounded datatypes actually generates a value based on some fixed range.

For Integer, random generates a value in the Int range. For Float and Double, random generates a floating point value in the range [0, 1).

This library does not provide Uniform instances for any unbounded datatypes.

Reproducibility

If you have two builds of a particular piece of code against this library, any deterministic function call should give the same result in the two builds if the builds are

compiled against the same major version of this library

on the same architecture (32-bit or 64-bit)

Pure and monadic pseudo-random number generators

Pseudo-random number generators come in two flavours: pure and monadic.

Pure pseudo-random number generators

These generators produce a new random value together with a new instance of the pseudo-random number generator. RandomGen defines the interface for pure pseudo-random number generators.
Pure pseudo-random number generators should implement split if they are splittable, that is, if there is an efficient method to turn one instance of the generator into two such that the pseudo-random numbers produced by the two resulting generators are not correlated. See [1] for some background on splittable pseudo-random generators.

Monadic pseudo-random number generators

These generators mutate their own state as they produce random values. They generally live in ST or IO or some transformer that implements PrimMonad. MonadRandom defines the interface for monadic pseudo-random number generators.

Pure pseudo-random number generators can be used in monadic code via the adapters Frozen, PrimGen and MutGen.

Frozen can be used in any state monad. With strict StateT there is even no performance overhead when compared to the RandomGen directly. But it is not safe to use it in presence of exceptions and resource allocation that requires cleanup.

PrimGen can be used in any PrimMonad if the pseudo-random number generator is also an instance of Prim. It will perform much faster than MutGen, but it can't be used in a concurrent setting.

MutGen can be used in any PrimMonad (this includes ST and IO) and can safely be shared between threads.

When to use which?

Use Frozen if your computation does not have side effects and results in a pure function.

Use PrimGen if the pseudo-random number generator implements Prim class and random value generation is intermixed with IO or ST.

Otherwise use MutGen. Whenever a MutGen is shared between threads, make sure there is not much contention for it, otherwise performance will suffer. For parallel random value generation it is best to split the generator and use a single Frozen or PrimGen per thread.

How to generate random values in monadic code

In monadic code, use the relevant Uniform and UniformRange instances to generate random values via uniform and uniformR, respectively.

As an example, rolls generates n random values of Word8 in the range [1, 6].

>>> :{
let rolls :: MonadRandom g s m => Int -> g s -> m [Word8]
    rolls n = replicateM n . uniformR (1, 6)
:}

Given a monadic pseudo-random number generator, you can run this probabilistic computation as follows:

>>> monadicGen <- MWC.create
>>> rolls 10 monadicGen :: IO [Word8]
[2,3,6,6,4,4,3,1,5,4]

Given a pure pseudo-random number generator, you can run it in an IO or ST context using runPrimGenIO or runPrimGenST and their variants, respectively. If the pseudo-random number generator does not implement Prim, you can also use runMutGenIO or runMutGenST and their variants.

>>> let pureGen = mkStdGen 42
>>> runGenM_ (IOGen pureGen) (rolls 10) :: IO [Word8]
[1,1,3,2,4,5,3,4,6,2]

How to generate random values in pure code

In pure code, use runGenState and its variants to extract the pure random value from a monadic computation based on a pure pseudo-random number generator.

>>> let pureGen = mkStdGen 42
>>> runGenState_ pureGen (rolls 10) :: [Word8]
[1,1,3,2,4,5,3,4,6,2]

How to implement RandomGen

Consider these points when writing a RandomGen instance for a given pure pseudo-random number generator:

If the pseudo-random number generator has a power-of-2 modulus, that is, it natively outputs 2^n bits of randomness for some n, implement genWord8, genWord16, genWord32 and genWord64. See below for more details.

If the pseudo-random number generator does not have a power-of-2 modulus, implement next and genRange. See below for more details.

If the pseudo-random number generator is splittable, implement split.

Additionally, implement Prim for the pseudo-random number generator if possible. This allows users to use the fast MutGen adapter with the pseudo-random number generator.

How to implement RandomGen for a pseudo-random number generator with power-of-2 modulus

Suppose you want to implement a permuted congruential generator.

>>> data PCGen = PCGen !Word64 !Word64

It produces a full Word32 of randomness per iteration.

>>> :{
let stepGen :: PCGen -> (Word32, PCGen)
    stepGen (PCGen state inc) = let
      newState = state * 6364136223846793005 + (inc .|. 1)
      xorShifted = fromIntegral (((state `shiftR` 18) `xor` state) `shiftR` 27) :: Word32
      rot = fromIntegral (state `shiftR` 59) :: Word32
      out = (xorShifted `shiftR` (fromIntegral rot)) .|. (xorShifted `shiftL` fromIntegral ((-rot) .&. 31))
      in (out, PCGen newState inc)
:}

>>> fst $ stepGen $ snd $ stepGen (PCGen 17 29)
3288430965

You can make it an instance of RandomGen as follows:

>>> :{
instance RandomGen PCGen where
  genWord32 = stepGen
  genWord64 g = (buildWord64 x y, g'')
    where
      (x, g') = stepGen g
      (y, g'') = stepGen g'
:}

This definition satisfies the compiler. However, the default implementations of genWord8 and genWord16 are geared towards backwards compatibility with RandomGen instances based on next and genRange. This means that they are not optimal for pseudo-random number generators with a power-of-2 modulo.

So let's implement a faster RandomGen instance for our pseudo-random number generator as follows:

>>> newtype PCGen' = PCGen' { unPCGen :: PCGen }
>>> let stepGen' = second PCGen' . stepGen . unPCGen
>>> :{
instance RandomGen PCGen' where
  genWord8 = first fromIntegral . stepGen'
  genWord16 = first fromIntegral . stepGen'
  genWord32 = stepGen'
  genWord64 g = (buildWord64 x y, g'')
    where
      (x, g') = stepGen' g
      (y, g'') = stepGen' g'
:}

How to implement RandomGen for a pseudo-random number generator without a power-of-2 modulus

We do not recommend you implement any new pseudo-random number generators without a power-of-2 modulus.

Pseudo-random number generators without a power-of-2 modulus perform significantly worse than pseudo-random number generators with a power-of-2 modulus with this library. This is because most functionality in this library is based on generating and transforming uniformly random machine words, and generating uniformly random machine words using a pseudo-random number generator without a power-of-2 modulus is expensive.

The pseudo-random number generator from L’Ecuyer (1988) natively generates an integer value in the range [1, 2147483562]. This is the generator used by this library before it was replaced by SplitMix in version 1.2.

>>> data LegacyGen = LegacyGen !Int32 !Int32
>>> :{
let legacyNext :: LegacyGen -> (Int, LegacyGen)
    legacyNext (LegacyGen s1 s2) = (fromIntegral z', LegacyGen s1'' s2'') where
      z' = if z < 1 then z + 2147483562 else z
      z = s1'' - s2''
      k = s1 `quot` 53668
      s1'  = 40014 * (s1 - k * 53668) - k * 12211
      s1'' = if s1' < 0 then s1' + 2147483563 else s1'
      k' = s2 `quot` 52774
      s2' = 40692 * (s2 - k' * 52774) - k' * 3791
      s2'' = if s2' < 0 then s2' + 2147483399 else s2'
:}

You can make it an instance of RandomGen as follows:

>>> :{
instance RandomGen LegacyGen where
  next = legacyNext
  genRange _ = (1, 2147483562)
:}

How to implement MonadRandom

Typically, a monadic pseudo-random number generator has facilities to save and restore its internal state in addition to generating random pseudo-random numbers.

Here is an example instance for the monadic pseudo-random number generator from the mwc-random package:

instance (s ~ PrimState m, PrimMonad m) => MonadRandom MWC.Gen s m where
  newtype Frozen MWC.Gen = Frozen { unFrozen :: MWC.Seed }
  thawGen = fmap MWC.restore unFrozen
  freezeGen = fmap Frozen . MWC.save
  uniformWord8 = MWC.uniform
  uniformWord16 = MWC.uniform
  uniformWord32 = MWC.uniform
  uniformWord64 = MWC.uniform
  uniformShortByteString n g = unsafeSTToPrim (genShortByteStringST n (MWC.uniform g))

Random number generators

class RandomGen g where

RandomGen is an interface to pure pseudo-random number generators.

StdGen is the default RandomGen instance provided by this library.

Minimal complete definition

split, (genWord32 | genWord64 | next, genRange)

Methods

next :: g -> (Int, g)

Deprecated: No longer used

Returns an Int that is uniformly distributed over the range returned by genRange (including both end points), and a new generator. Using next is inefficient as all operations go via Integer. See here for more details. It is thus deprecated.

genWord8 :: g -> (Word8, g)

Returns a Word8 that is uniformly distributed over the entire Word8 range.

Since: 1.2

genWord16 :: g -> (Word16, g)

Returns a Word16 that is uniformly distributed over the entire Word16 range.

Since: 1.2

genWord32 :: g -> (Word32, g)

Returns a Word32 that is uniformly distributed over the entire Word32 range.

Since: 1.2

genWord64 :: g -> (Word64, g)

Returns a Word64 that is uniformly distributed over the entire Word64 range.

Since: 1.2

genWord32R :: Word32 -> g -> (Word32, g)

genWord32R upperBound g returns a Word32 that is uniformly distributed over the range [0, upperBound].

Since: 1.2

genWord64R :: Word64 -> g -> (Word64, g)

genWord64R upperBound g returns a Word64 that is uniformly distributed over the range [0, upperBound].

Since: 1.2

genShortByteString :: Int -> g -> (ShortByteString, g)

genShortByteString n g returns a ShortByteString of length n filled with random bytes.

Since: 1.2

genRange :: g -> (Int, Int)

Deprecated: No longer used

Yields the range of values returned by next.

It is required that:

If (a, b) = genRange g, then a < b.

genRange must not examine its argument so the value it returns is determined only by the instance of RandomGen.

The default definition spans the full range of Int.

split :: g -> (g, g)

Returns two distinct pseudo-random number generators.

Implementations should take care to ensure that the resulting generators are not correlated. Some pseudo-random number generators are not splittable. In that case, the split implementation should fail with a descriptive error message.

class Monad m => MonadRandom g s m | m -> s where

MonadRandom is an interface to monadic pseudo-random number generators.

Minimal complete definition

freezeGen, thawGen, (uniformWord32 | uniformWord64)

Associated Types

data Frozen g :: *

Represents the state of the pseudo-random number generator for use with thawGen and freezeGen.

Since: 1.2

Methods

thawGen :: Frozen g -> m (g s)

Restores the pseudo-random number generator from its Frozen representation.

Since: 1.2

freezeGen :: g s -> m (Frozen g)

Saves the state of the pseudo-random number generator to its Frozen representation.

Since: 1.2

uniformWord32R :: Word32 -> g s -> m Word32

uniformWord32R upperBound g generates a Word32 that is uniformly distributed over the range [0, upperBound].

Since: 1.2

uniformWord64R :: Word64 -> g s -> m Word64

uniformWord64R upperBound g generates a Word64 that is uniformly distributed over the range [0, upperBound].

Since: 1.2

uniformWord8 :: g s -> m Word8

Generates a Word8 that is uniformly distributed over the entire Word8 range.

The default implementation extracts a Word8 from uniformWord32.

Since: 1.2

uniformWord16 :: g s -> m Word16

Generates a Word16 that is uniformly distributed over the entire Word16 range.

The default implementation extracts a Word16 from uniformWord32.

Since: 1.2

uniformWord32 :: g s -> m Word32

Generates a Word32 that is uniformly distributed over the entire Word32 range.

The default implementation extracts a Word32 from uniformWord64.

Since: 1.2

uniformWord64 :: g s -> m Word64

Generates a Word64 that is uniformly distributed over the entire Word64 range.

The default implementation combines two Word32 from uniformWord32 into one Word64.

Since: 1.2

uniformShortByteString :: Int -> g s -> m ShortByteString

uniformShortByteString n g generates a ShortByteString of length n filled with random bytes.

Since: 1.2

uniformShortByteString :: MonadIO m => Int -> g s -> m ShortByteString

uniformShortByteString n g generates a ShortByteString of length n filled with random bytes.

Since: 1.2

runGenM :: MonadRandom g s m => Frozen g -> (g s -> m a) -> m (a, Frozen g)

Runs a mutable pseudo-random number generator from its Frozen state.

>>> import Data.Int (Int8)
>>> runGenM (IOGen (mkStdGen 217)) (`uniformListM` 5) :: IO ([Int8], Frozen (IOGen StdGen))
([-74,37,-50,-2,3],IOGen {unIOGen = SMGen 4273268533320920145 15251669095119325999})

Since: 1.2

runGenM_ :: MonadRandom g s m => Frozen g -> (g s -> m a) -> m a

Same as runGenM, but only returns the generated value.

Since: 1.2

class (RandomGen r, MonadRandom (g r) s m) => RandomGenM g r s m where

Interface to operations on RandomGen wrappers like Frozen and Frozen.

Since: 1.2

Methods

applyRandomGenM :: (r -> (a, r)) -> g r s -> m a

splitRandomGenM :: RandomGenM g r s m => g r s -> m r

Splits a pseudo-random number generator into two. Overwrites the mutable wrapper with one of the resulting generators and returns the other.

Since: 1.2

Standard random number generators

type StdGen = SMGen

The default pseudo-random number generator.

mkStdGen :: Int -> StdGen

The function mkStdGen provides an alternative way of producing an initial generator, by mapping an Int into a generator. Again, distinct arguments should be likely to produce distinct generators.

Stateful interface for pure generators

Based on StateT

data PureGen g s

Opaque data type that carries the type of a pure pseudo-random number generator.

Since: 1.2

splitGen :: (MonadState g m, RandomGen g) => m g

Splits a pseudo-random number generator into two. Updates the state with one of the resulting generators and returns the other.

Since: 1.2

genRandom :: (RandomGen g, Random a, MonadState g m) => PureGen g g -> m a

Generates a random value in a state monad.

Since: 1.2

runGenState :: RandomGen g => g -> (PureGen g g -> State g a) -> (a, g)

Runs a monadic generating action in the State monad using a pure pseudo-random number generator.

Since: 1.2

runGenState_ :: RandomGen g => g -> (PureGen g g -> State g a) -> a

Runs a monadic generating action in the State monad using a pure pseudo-random number generator. Returns only the resulting random value.

Since: 1.2

runGenStateT :: RandomGen g => g -> (PureGen g g -> StateT g m a) -> m (a, g)

Runs a monadic generating action in the StateT monad using a pure pseudo-random number generator.

Since: 1.2

runGenStateT_ :: (RandomGen g, Functor f) => g -> (PureGen g g -> StateT g f a) -> f a

Runs a monadic generating action in the StateT monad using a pure pseudo-random number generator. Returns only the resulting random value.

Since: 1.2

runPureGenST :: RandomGen g => g -> (forall s. PureGen g g -> StateT g (ST s) a) -> (a, g)

Runs a monadic generating action in the ST monad using a pure pseudo-random number generator.

Since: 1.2

Mutable generators

AtomicGen

data AtomicGen g s

This is a wrapper around pure generator that can be used in a monadic environment. It is safe in presence of exceptions and concurrency since all operations are performed atomically.

Since: 1.2

applyAtomicGen :: MonadIO m => (g -> (a, g)) -> AtomicGen g RealWorld -> m a

Atomically applies a pure operation to the wrapped pseudo-random number generator.

Since: 1.2

IOGen

data IOGen g s

This is a wrapper around an IORef that holds a pure generator. Because of extra pointer indirection it will be slightly slower than if Frozen is being used, but faster than Frozen wrapper, since atomic modification is not being used with Frozen. Which also means that it is not safe in a concurrent setting.

Both Frozen and Frozen are necessary when generation of random values happens in IO and especially when dealing with exception handling and resource allocation, which is where StateT should never be used. For example writing a random number of bytes into a temporary file:

>>> import UnliftIO.Temporary (withSystemTempFile)
>>> import Data.ByteString (hPutStr)
>>> let ioGen g = withSystemTempFile "foo.bin" $ \_ h -> uniformR (0, 100) g >>= flip uniformByteString g >>= hPutStr h

and then run it:

>>> runGenM_ (IOGen (mkStdGen 1729)) ioGen

Since: 1.2

applyIOGen :: MonadIO m => (g -> (a, g)) -> IOGen g RealWorld -> m a

Applies a pure operation to the wrapped pseudo-random number generator.

Since: 1.2

STGen

data STGen g s

This is a wrapper wround an STRef that holds a pure generator. Because of extra pointer indirection it will be slightly slower than if Frozen is being used.

Since: 1.2

applySTGen :: (g -> (a, g)) -> STGen g s -> ST s a

Applies a pure operation to the wrapped pseudo-random number generator.

Since: 1.2

runSTGen :: RandomGen g => g -> (forall s. STGen g s -> ST s a) -> (a, g)

Runs a monadic generating action in the ST monad using a pure pseudo-random number generator.

Since: 1.2

runSTGen_ :: RandomGen g => g -> (forall s. STGen g s -> ST s a) -> a

Runs a monadic generating action in the ST monad using a pure pseudo-random number generator. Returns only the resulting random value.

Since: 1.2

The global random number generator

There is a single, implicit, global random number generator of type StdGen, held in some global variable maintained by the IO monad. It is initialised automatically in some system-dependent fashion, for example, by using the time of day, or Linux's kernel random number generator. To get deterministic behaviour, use setStdGen.

getStdRandom :: (StdGen -> (a, StdGen)) -> IO a

Uses the supplied function to get a value from the current global random generator, and updates the global generator with the new generator returned by the function. For example, rollDice gets a random integer between 1 and 6:

 rollDice :: IO Int
 rollDice = getStdRandom (randomR (1,6))

getStdGen :: IO StdGen

Gets the global random number generator.

setStdGen :: StdGen -> IO ()

Sets the global random number generator.

newStdGen :: IO StdGen

Applies split to the current global random generator, updates it with one of the results, and returns the other.

Random values of various types

random has two type classes for generation of random numbers: Uniform and UniformRange. One for generating every possible value and another for generating every value in range. In other libraries this functionality frequently bundled into single type class but here we have two type classes because there're types which could have instance for one type class but not the other.

For example: Integer, Float, Double have instance for UniformRange but there's no way to define Uniform.

Conversely there're types where Uniform instance is possible while UniformRange is not. One example is tuples: (a,b). While Uniform instance is straightforward it's not clear how to define UniformRange. We could try to generate values that a <= x <= b But to do that we need to know number of elements of tuple's second type parameter b which we don't have.

Or type could have no order at all. Take for example angle. Defining Uniform instance is again straghtforward: just generate value in [0,2π) range. But for any two pair of angles there're two ranges: clockwise and counterclockwise.

class Uniform a where

Generate every possible value for data type with equal probability.

Since: 1.2

Methods

uniform :: MonadRandom g s m => g s -> m a

uniformListM :: (MonadRandom g s m, Uniform a) => g s -> Int -> m [a]

Generates a list of random values.

Since: 1.2

class UniformRange a where

Generate every value in provided inclusive range with equal probability. So uniformR (1,4) should generate values from set [1,2,3,4]. Inclusive range is used to allow to express any interval for fixed-size ints, enumerations etc.

Additionally in order to make function always defined order of elements in range shouldn't matter and following law should hold:

uniformR (a,b) = uniform (b,a)

Since: 1.2

Methods

uniformR :: MonadRandom g s m => (a, a) -> g s -> m a

class Random a where

With a source of random number supply in hand, the Random class allows the programmer to extract random values of a variety of types.

Minimal complete definition: randomR and random.

Minimal complete definition

randomM

Methods

randomR :: RandomGen g => (a, a) -> g -> (a, g)

Takes a range (lo,hi) and a random number generator g, and returns a random value uniformly distributed over the closed interval [lo,hi], together with a new generator. It is unspecified what happens if lo>hi. For continuous types there is no requirement that the values lo and hi are ever produced, but they may be, depending on the implementation and the interval.

randomR :: (RandomGen g, UniformRange a) => (a, a) -> g -> (a, g)

Takes a range (lo,hi) and a random number generator g, and returns a random value uniformly distributed over the closed interval [lo,hi], together with a new generator. It is unspecified what happens if lo>hi. For continuous types there is no requirement that the values lo and hi are ever produced, but they may be, depending on the implementation and the interval.

random :: RandomGen g => g -> (a, g)

The same as randomR, but using a default range determined by the type:

For bounded types (instances of Bounded, such as Char), the range is normally the whole type.

For fractional types, the range is normally the semi-closed interval [0,1).

For Integer, the range is (arbitrarily) the range of Int.

randomM :: MonadRandom g s m => g s -> m a

randomRs :: RandomGen g => (a, a) -> g -> [a]

Plural variant of randomR, producing an infinite list of random values instead of returning a new generator.

randoms :: RandomGen g => g -> [a]

Plural variant of random, producing an infinite list of random values instead of returning a new generator.

randomRIO :: (a, a) -> IO a

Deprecated: In favor of uniformR

A variant of randomR that uses the global random number generator (see System.Random).

randomIO :: IO a

Deprecated: In favor of uniformR

A variant of random that uses the global random number generator (see System.Random).

Generators for sequences of bytes

genShortByteStringWith :: MonadIO m => Int -> m Word64 -> m ShortByteString

Efficiently generates a sequence of random bytes in a platform independent manner. The allocated memory is be pinned, so it is safe to use with FFI calls.

Since: 1.2

genShortByteStringST :: Int -> ST s Word64 -> ST s ShortByteString

Same as genShortByteStringWith, but runs in ST.

Since: 1.2

uniformByteString :: MonadRandom g s m => Int -> g s -> m ByteString

Generates a random ByteString of the specified size.

Since: 1.2

genByteString :: RandomGen g => Int -> g -> (ByteString, g)

Generates a ByteString of the specified size using a pure pseudo-random number generator. See uniformByteString for the monadic version.

Since: 1.2

References

Guy L. Steele, Jr., Doug Lea, and Christine H. Flood. 2014. Fast splittable pseudorandom number generators. In Proceedings of the 2014 ACM International Conference on Object Oriented Programming Systems Languages & Applications (OOPSLA '14). ACM, New York, NY, USA, 453-472. DOI: https://doi.org/10.1145/2660193.2660195