System.Random

Description

This library deals with the common task of pseudo-random number generation.

Introduction

This module provides type classes and instances for the following concepts:

Pure pseudo-random number generators

RandomGen is an interface to pure pseudo-random number generators.
StdGen, the standard pseudo-random number generator provided in this library, is an instance of RandomGen. It uses the SplitMix implementation provided by the splitmix package. Programmers may, of course, supply their own instances of RandomGen.

Usage

In pure code, use uniform and uniformR to generate pseudo-random values with a pure pseudo-random number generator like StdGen.

>>> :{
let rolls :: RandomGen g => Int -> g -> [Word]
    rolls n = take n . unfoldr (Just . uniformR (1, 6))
    pureGen = mkStdGen 137
in
    rolls 10 pureGen :: [Word]
:}
[4,2,6,1,6,6,5,1,1,5]

To run use a monadic pseudo-random computation in pure code with a pure pseudo-random number generator, use runStateGen and its variants.

>>> :{
let rollsM :: StatefulGen g m => Int -> g -> m [Word]
    rollsM n = replicateM n . uniformRM (1, 6)
    pureGen = mkStdGen 137
in
    runStateGen_ pureGen (rollsM 10) :: [Word]
:}
[4,2,6,1,6,6,5,1,1,5]

Pure number generator interface

Pseudo-random number generators come in two flavours: pure and monadic.

RandomGen: pure pseudo-random number generators

These generators produce a new pseudo-random value together with a new instance of the pseudo-random number generator.
Pure pseudo-random number generators should implement split if they are splittable, that is, if there is an efficient method to turn one generator into two. The pseudo-random numbers produced by the two resulting generators should not be correlated. See [1] for some background on splittable pseudo-random generators.

StatefulGen: monadic pseudo-random number generators

See System.Random.Stateful module

class RandomGen g where

RandomGen is an interface to pure pseudo-random number generators.

StdGen is the standard RandomGen instance provided by this library.

Since: 1.0.0

Minimal complete definition

split, (genWord32 | genWord64 | next, genRange)

Methods

next :: g -> (Int, g)

Deprecated: No longer used

Returns an Int that is uniformly distributed over the range returned by genRange (including both end points), and a new generator. Using next is inefficient as all operations go via Integer. See here for more details. It is thus deprecated.

Since: 1.0.0

genWord8 :: g -> (Word8, g)

Returns a Word8 that is uniformly distributed over the entire Word8 range.

Since: 1.2.0

genWord16 :: g -> (Word16, g)

Returns a Word16 that is uniformly distributed over the entire Word16 range.

Since: 1.2.0

genWord32 :: g -> (Word32, g)

Returns a Word32 that is uniformly distributed over the entire Word32 range.

Since: 1.2.0

genWord64 :: g -> (Word64, g)

Returns a Word64 that is uniformly distributed over the entire Word64 range.

Since: 1.2.0

genWord32R :: Word32 -> g -> (Word32, g)

genWord32R upperBound g returns a Word32 that is uniformly distributed over the range [0, upperBound].

Since: 1.2.0

genWord64R :: Word64 -> g -> (Word64, g)

genWord64R upperBound g returns a Word64 that is uniformly distributed over the range [0, upperBound].

Since: 1.2.0

genShortByteString :: Int -> g -> (ShortByteString, g)

genShortByteString n g returns a ShortByteString of length n filled with pseudo-random bytes.

Since: 1.2.0

genRange :: g -> (Int, Int)

Deprecated: No longer used

Yields the range of values returned by next.

It is required that:

If (a, b) = genRange g, then a < b.

genRange must not examine its argument so the value it returns is determined only by the instance of RandomGen.

The default definition spans the full range of Int.

Since: 1.0.0

split :: g -> (g, g)

Returns two distinct pseudo-random number generators.

Implementations should take care to ensure that the resulting generators are not correlated. Some pseudo-random number generators are not splittable. In that case, the split implementation should fail with a descriptive error message.

Since: 1.0.0

uniform :: (RandomGen g, Uniform a) => g -> (a, g)

Generates a value uniformly distributed over all possible values of that type.

This is a pure version of uniformM.

Examples

Expand

>>> import System.Random
>>> let pureGen = mkStdGen 137
>>> uniform pureGen :: (Bool, StdGen)
(True,StdGen {unStdGen = SMGen 11285859549637045894 7641485672361121627})

Since: 1.2.0

uniformR :: (RandomGen g, UniformRange a) => (a, a) -> g -> (a, g)

Generates a value uniformly distributed over the provided range, which is interpreted as inclusive in the lower and upper bound.

uniformR (1 :: Int, 4 :: Int) generates values uniformly from the set \(\{1,2,3,4\}\)

uniformR (1 :: Float, 4 :: Float) generates values uniformly from the set \(\{x\;|\;1 \le x \le 4\}\)

The following law should hold to make the function always defined:

uniformR (a, b) = uniformR (b, a)

This is a pure version of uniformRM.

Examples

Expand

>>> import System.Random
>>> let pureGen = mkStdGen 137
>>> uniformR (1 :: Int, 4 :: Int) pureGen
(4,StdGen {unStdGen = SMGen 11285859549637045894 7641485672361121627})

Since: 1.2.0

genByteString :: RandomGen g => Int -> g -> (ByteString, g)

Generates a ByteString of the specified size using a pure pseudo-random number generator. See uniformByteStringM for the monadic version.

Examples

Expand

>>> import System.Random
>>> import Data.ByteString
>>> let pureGen = mkStdGen 137
>>> unpack . fst . genByteString 10 $ pureGen
[51,123,251,37,49,167,90,109,1,4]

Since: 1.2.0

class Random a where

The class of types for which random values can be generated. Most instances of Random will produce values that are uniformly distributed on the full range, but for those types without a well-defined "full range" some sensible default subrange will be selected.

Random exists primarily for backwards compatibility with version 1.1 of this library. In new code, use the better specified Uniform and UniformRange instead.

Since: 1.0.0

Minimal complete definition

Nothing

Methods

randomR :: RandomGen g => (a, a) -> g -> (a, g)

Takes a range (lo,hi) and a pseudo-random number generator g, and returns a pseudo-random value uniformly distributed over the closed interval [lo,hi], together with a new generator. It is unspecified what happens if lo>hi, but usually the values will simply get swapped.

>>> let gen = mkStdGen 2021
>>> fst $ randomR ('a', 'z') gen
't'
>>> fst $ randomR ('z', 'a') gen
't'

For continuous types there is no requirement that the values lo and hi are ever produced, but they may be, depending on the implementation and the interval.

There is no requirement to follow the Ord instance and the concept of range can be defined on per type basis. For example product types will treat their values independently:

>>> fst $ randomR (('a', 5.0), ('z', 10.0)) $ mkStdGen 2021
('t',6.240232662366563)

In case when a lawful range is desired uniformR should be used instead.

Since: 1.0.0

default randomR :: (RandomGen g, UniformRange a) => (a, a) -> g -> (a, g)

random :: RandomGen g => g -> (a, g)

The same as randomR, but using a default range determined by the type:

For bounded types (instances of Bounded, such as Char), the range is normally the whole type.

For floating point types, the range is normally the closed interval [0,1].

For Integer, the range is (arbitrarily) the range of Int.

Since: 1.0.0

default random :: (RandomGen g, Uniform a) => g -> (a, g)

randomRs :: RandomGen g => (a, a) -> g -> [a]

Plural variant of randomR, producing an infinite list of pseudo-random values instead of returning a new generator.

Since: 1.0.0

randoms :: RandomGen g => g -> [a]

Plural variant of random, producing an infinite list of pseudo-random values instead of returning a new generator.

Since: 1.0.0

class Uniform a

The class of types for which a uniformly distributed value can be drawn from all possible values of the type.

Since: 1.2.0

class UniformRange a

The class of types for which a uniformly distributed value can be drawn from a range.

Since: 1.2.0

Minimal complete definition

uniformRM

class Finite a

A type class for data with a finite number of inhabitants. This type class is used in default implementations of Uniform.

Users are not supposed to write instances of Finite manually. There is a default implementation in terms of Generic instead.

>>> :set -XDeriveGeneric -XDeriveAnyClass
>>> import GHC.Generics (Generic)
>>> data MyBool = MyTrue | MyFalse deriving (Generic, Finite)
>>> data Action = Code MyBool | Eat (Maybe Bool) | Sleep deriving (Generic, Finite)

Standard pseudo-random number generator

data StdGen

The standard pseudo-random number generator.

mkStdGen :: Int -> StdGen

Constructs a StdGen deterministically.

initStdGen :: MonadIO m => m StdGen

Initialize StdGen using system entropy (i.e. /dev/urandom) when it is available, while falling back on using system time as the seed.

Since: 1.2.1

Global standard pseudo-random number generator

There is a single, implicit, global pseudo-random number generator of type StdGen, held in a global mutable variable that can be manipulated from within the IO monad. It is also available as globalStdGen, therefore it is recommended to use the new System.Random.Stateful interface to explicitly operate on the global pseudo-random number generator.

It is initialised with initStdGen, although it is possible to override its value with setStdGen. All operations on the global pseudo-random number generator are thread safe, however in presence of concurrency they are naturally become non-deterministic. Moreover, relying on the global mutable state makes it hard to know which of the dependent libraries are using it as well, making it unpredictable in the local context. Precisely of this reason, the global pseudo-random number generator is only suitable for uses in applications, test suites, etc. and is advised against in development of reusable libraries.

It is also important to note that either using StdGen with pure functions from other sections of this module or by relying on runStateGen from stateful interface does not only give us deterministic behaviour without requiring IO, but it is also more efficient.

getStdRandom :: MonadIO m => (StdGen -> (a, StdGen)) -> m a

Uses the supplied function to get a value from the current global random generator, and updates the global generator with the new generator returned by the function. For example, rollDice produces a pseudo-random integer between 1 and 6:

>>> rollDice = getStdRandom (randomR (1, 6))
>>> replicateM 10 (rollDice :: IO Int)
[5,6,6,1,1,6,4,2,4,1]

This is an outdated function and it is recommended to switch to its equivalent applyAtomicGen instead, possibly with the globalStdGen if relying on the global state is acceptable.

>>> import System.Random.Stateful
>>> rollDice = applyAtomicGen (uniformR (1, 6)) globalStdGen
>>> replicateM 10 (rollDice :: IO Int)
[4,6,1,1,4,4,3,2,1,2]

Since: 1.0.0

getStdGen :: MonadIO m => m StdGen

Gets the global pseudo-random number generator. Extracts the contents of globalStdGen

Since: 1.0.0

setStdGen :: MonadIO m => StdGen -> m ()

Sets the global pseudo-random number generator. Overwrites the contents of globalStdGen

Since: 1.0.0

newStdGen :: MonadIO m => m StdGen

Applies split to the current global pseudo-random generator globalStdGen, updates it with one of the results, and returns the other.

Since: 1.0.0

randomIO :: (Random a, MonadIO m) => m a

A variant of randomM that uses the global pseudo-random number generator globalStdGen.

>>> import Data.Int
>>> randomIO :: IO Int32
-1580093805

This function is equivalent to getStdRandom random and is included in this interface for historical reasons and backwards compatibility. It is recommended to use uniformM instead, possibly with the globalStdGen if relying on the global state is acceptable.

>>> import System.Random.Stateful
>>> uniformM globalStdGen :: IO Int32
-1649127057

Since: 1.0.0

randomRIO :: (Random a, MonadIO m) => (a, a) -> m a

A variant of randomRM that uses the global pseudo-random number generator globalStdGen

>>> randomRIO (2020, 2100) :: IO Int
2040

Similar to randomIO, this function is equivalent to getStdRandom randomR and is included in this interface for historical reasons and backwards compatibility. It is recommended to use uniformRM instead, possibly with the globalStdGen if relying on the global state is acceptable.

>>> import System.Random.Stateful
>>> uniformRM (2020, 2100) globalStdGen :: IO Int
2079

Since: 1.0.0

Compatibility and reproducibility

Backwards compatibility and deprecations

Version 1.2 mostly maintains backwards compatibility with version 1.1. This has a few consequences users should be aware of:

The type class Random is only provided for backwards compatibility. New code should use Uniform and UniformRange instead.

The methods next and genRange in RandomGen are deprecated and only provided for backwards compatibility. New instances of RandomGen should implement word-based methods instead. See below for more information about how to write a RandomGen instance.

This library provides instances for Random for some unbounded types for backwards compatibility. For an unbounded type, there is no way to generate a value with uniform probability out of its entire domain, so the random implementation for unbounded types actually generates a value based on some fixed range.

For Integer, random generates a value in the Int range. For Float and Double, random generates a floating point value in the range [0, 1).

This library does not provide Uniform instances for any unbounded types.

Reproducibility

If you have two builds of a particular piece of code against this library, any deterministic function call should give the same result in the two builds if the builds are

compiled against the same major version of this library

on the same architecture (32-bit or 64-bit)

Notes for pseudo-random number generator implementors

How to implement RandomGen

Consider these points when writing a RandomGen instance for a given pure pseudo-random number generator:

If the pseudo-random number generator has a power-of-2 modulus, that is, it natively outputs 2^n bits of randomness for some n, implement genWord8, genWord16, genWord32 and genWord64. See below for more details.

If the pseudo-random number generator does not have a power-of-2 modulus, implement next and genRange. See below for more details.

If the pseudo-random number generator is splittable, implement split. If there is no suitable implementation, split should fail with a helpful error message.

How to implement RandomGen for a pseudo-random number generator with power-of-2 modulus

Suppose you want to implement a permuted congruential generator.

>>> data PCGen = PCGen !Word64 !Word64

It produces a full Word32 of randomness per iteration.

>>> import Data.Bits
>>> :{
let stepGen :: PCGen -> (Word32, PCGen)
    stepGen (PCGen state inc) = let
      newState = state * 6364136223846793005 + (inc .|. 1)
      xorShifted = fromIntegral (((state `shiftR` 18) `xor` state) `shiftR` 27) :: Word32
      rot = fromIntegral (state `shiftR` 59) :: Word32
      out = (xorShifted `shiftR` (fromIntegral rot)) .|. (xorShifted `shiftL` fromIntegral ((-rot) .&. 31))
      in (out, PCGen newState inc)
:}

>>> fst $ stepGen $ snd $ stepGen (PCGen 17 29)
3288430965

You can make it an instance of RandomGen as follows:

>>> :{
instance RandomGen PCGen where
  genWord32 = stepGen
  split _ = error "PCG is not splittable"
:}

How to implement RandomGen for a pseudo-random number generator without a power-of-2 modulus

We do not recommend you implement any new pseudo-random number generators without a power-of-2 modulus.

Pseudo-random number generators without a power-of-2 modulus perform significantly worse than pseudo-random number generators with a power-of-2 modulus with this library. This is because most functionality in this library is based on generating and transforming uniformly pseudo-random machine words, and generating uniformly pseudo-random machine words using a pseudo-random number generator without a power-of-2 modulus is expensive.

The pseudo-random number generator from L’Ecuyer (1988) natively generates an integer value in the range [1, 2147483562]. This is the generator used by this library before it was replaced by SplitMix in version 1.2.

>>> data LegacyGen = LegacyGen !Int32 !Int32
>>> :{
let legacyNext :: LegacyGen -> (Int, LegacyGen)
    legacyNext (LegacyGen s1 s2) = (fromIntegral z', LegacyGen s1'' s2'') where
      z' = if z < 1 then z + 2147483562 else z
      z = s1'' - s2''
      k = s1 `quot` 53668
      s1'  = 40014 * (s1 - k * 53668) - k * 12211
      s1'' = if s1' < 0 then s1' + 2147483563 else s1'
      k' = s2 `quot` 52774
      s2' = 40692 * (s2 - k' * 52774) - k' * 3791
      s2'' = if s2' < 0 then s2' + 2147483399 else s2'
:}

You can make it an instance of RandomGen as follows:

>>> :{
instance RandomGen LegacyGen where
  next = legacyNext
  genRange _ = (1, 2147483562)
  split _ = error "Not implemented"
:}

References

Guy L. Steele, Jr., Doug Lea, and Christine H. Flood. 2014. Fast splittable pseudorandom number generators. In Proceedings of the 2014 ACM International Conference on Object Oriented Programming Systems Languages & Applications (OOPSLA '14). ACM, New York, NY, USA, 453-472. DOI: https://doi.org/10.1145/2660193.2660195