System.Random.Stateful

Description

This library deals with the common task of pseudo-random number generation.

Pure Random Generator

module System.Random

Monadic Random Generator

This module provides type classes and instances for the following concepts:

Monadic pseudo-random number generators

StatefulGen is an interface to monadic pseudo-random number generators.

Monadic adapters

StateGenM, AtomicGenM, IOGenM, STGenM and TGenM turn a RandomGen instance into a StatefulGen instance.

Drawing from a range

UniformRange is used to generate a value of a type uniformly within a range.
This library provides instances of UniformRange for many common numeric types.

Drawing from the entire domain of a type

Uniform is used to generate a value of a type uniformly over all possible values of that type.
This library provides instances of Uniform for many common bounded numeric types.

Usage

In monadic code, use the relevant Uniform and UniformRange instances to generate pseudo-random values via uniformM and uniformRM, respectively.

As an example, rollsM generates n pseudo-random values of Word in the range [1, 6] in a StatefulGen context; given a monadic pseudo-random number generator, you can run this probabilistic computation as follows:

>>> :{
let rollsM :: StatefulGen g m => Int -> g -> m [Word]
    rollsM n = replicateM n . uniformRM (1, 6)
in do
    monadicGen <- MWC.create
    rollsM 10 monadicGen :: IO [Word]
:}
[3,4,3,1,4,6,1,6,1,4]

Given a pure pseudo-random number generator, you can run the monadic pseudo-random number computation rollsM in an IO or ST context by applying a monadic adapter like AtomicGenM, IOGenM or STGenM (see monadic-adapters) to the pure pseudo-random number generator.

>>> :{
let rollsM :: StatefulGen g m => Int -> g -> m [Word]
    rollsM n = replicateM n . uniformRM (1, 6)
    pureGen = mkStdGen 42
in
    newIOGenM pureGen >>= rollsM 10 :: IO [Word]
:}
[1,1,3,2,4,5,3,4,6,2]

Mutable pseudo-random number generator interfaces

Pseudo-random number generators come in two flavours: pure and monadic.

RandomGen: pure pseudo-random number generators

See System.Random module.

StatefulGen: monadic pseudo-random number generators

These generators mutate their own state as they produce pseudo-random values. They generally live in ST or IO or some transformer that implements PrimMonad.

class Monad m => StatefulGen g m where

StatefulGen is an interface to monadic pseudo-random number generators.

Since: 1.2.0

Minimal complete definition

(uniformWord32 | uniformWord64)

Methods

uniformWord32R :: Word32 -> g -> m Word32

uniformWord32R upperBound g generates a Word32 that is uniformly distributed over the range [0, upperBound].

Since: 1.2.0

uniformWord64R :: Word64 -> g -> m Word64

uniformWord64R upperBound g generates a Word64 that is uniformly distributed over the range [0, upperBound].

Since: 1.2.0

uniformWord8 :: g -> m Word8

Generates a Word8 that is uniformly distributed over the entire Word8 range.

The default implementation extracts a Word8 from uniformWord32.

Since: 1.2.0

uniformWord16 :: g -> m Word16

Generates a Word16 that is uniformly distributed over the entire Word16 range.

The default implementation extracts a Word16 from uniformWord32.

Since: 1.2.0

uniformWord32 :: g -> m Word32

Generates a Word32 that is uniformly distributed over the entire Word32 range.

The default implementation extracts a Word32 from uniformWord64.

Since: 1.2.0

uniformWord64 :: g -> m Word64

Generates a Word64 that is uniformly distributed over the entire Word64 range.

The default implementation combines two Word32 from uniformWord32 into one Word64.

Since: 1.2.0

uniformShortByteString :: Int -> g -> m ShortByteString

uniformShortByteString n g generates a ShortByteString of length n filled with pseudo-random bytes.

Since: 1.2.0

default uniformShortByteString :: MonadIO m => Int -> g -> m ShortByteString

class StatefulGen (MutableGen f m) m => FrozenGen f m where

This class is designed for stateful pseudo-random number generators that can be saved as and restored from an immutable data type.

Since: 1.2.0

Associated Types

type MutableGen f m = (g :: Type) | g -> f

Represents the state of the pseudo-random number generator for use with thawGen and freezeGen.

Since: 1.2.0

Methods

freezeGen :: MutableGen f m -> m f

Saves the state of the pseudo-random number generator as a frozen seed.

Since: 1.2.0

thawGen :: f -> m (MutableGen f m)

Restores the pseudo-random number generator from its frozen seed.

Since: 1.2.0

class (RandomGen r, StatefulGen g m) => RandomGenM g r m | g -> r where

Interface to operations on RandomGen wrappers like IOGenM and StateGenM.

Since: 1.2.0

Methods

applyRandomGenM :: (r -> (a, r)) -> g -> m a

withMutableGen :: FrozenGen f m => f -> (MutableGen f m -> m a) -> m (a, f)

Runs a mutable pseudo-random number generator from its FrozenGen state.

Examples

Expand

>>> import Data.Int (Int8)
>>> withMutableGen (IOGen (mkStdGen 217)) (uniformListM 5) :: IO ([Int8], IOGen StdGen)
([-74,37,-50,-2,3],IOGen {unIOGen = StdGen {unStdGen = SMGen 4273268533320920145 15251669095119325999}})

Since: 1.2.0

withMutableGen_ :: FrozenGen f m => f -> (MutableGen f m -> m a) -> m a

Same as withMutableGen, but only returns the generated value.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> withMutableGen_ (IOGen pureGen) (uniformRM (1 :: Int, 6 :: Int))
4

Since: 1.2.0

randomM :: (RandomGenM g r m, Random a) => g -> m a

Generates a pseudo-random value using monadic interface and Random instance.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> g <- newIOGenM pureGen
>>> randomM g :: IO Double
0.5728354935654512

Since: 1.2.0

randomRM :: (RandomGenM g r m, Random a) => (a, a) -> g -> m a

Generates a pseudo-random value using monadic interface and Random instance.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> g <- newIOGenM pureGen
>>> randomRM (1, 100) g :: IO Int
52

Since: 1.2.0

splitGenM :: RandomGenM g r m => g -> m r

Splits a pseudo-random number generator into two. Overwrites the mutable wrapper with one of the resulting generators and returns the other.

Since: 1.2.0

Monadic adapters for pure pseudo-random number generators

Pure pseudo-random number generators can be used in monadic code via the adapters StateGenM, AtomicGenM, IOGenM, STGenM and TGenM

StateGenM can be used in any state monad. With strict StateT there is no performance overhead compared to using the RandomGen instance directly. StateGenM is not safe to use in the presence of exceptions and concurrency.

AtomicGenM is safe in the presence of exceptions and concurrency since it performs all actions atomically.

IOGenM is a wrapper around an IORef that holds a pure generator. IOGenM is safe in the presence of exceptions, but not concurrency.

STGenM is a wrapper around an STRef that holds a pure generator. STGenM is safe in the presence of exceptions, but not concurrency.

TGenM is a wrapper around a TVar that holds a pure generator. TGenM can be used in a software transactional memory monad STM. It is not as performant as AtomicGenM, but it can provide stronger guarantees in a concurrent setting.

Pure adapter

newtype StateGen g

Wrapper for pure state gen, which acts as an immutable seed for the corresponding stateful generator StateGenM

Since: 1.2.0

Constructors

StateGen

Fields

unStateGen :: g

data StateGenM g

Opaque data type that carries the type of a pure pseudo-random number generator.

Since: 1.2.0

Constructors

StateGenM

runStateGen :: RandomGen g => g -> (StateGenM g -> State g a) -> (a, g)

Runs a monadic generating action in the State monad using a pure pseudo-random number generator.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> runStateGen pureGen randomM :: (Int, StdGen)
(7879794327570578227,StdGen {unStdGen = SMGen 11285859549637045894 7641485672361121627})

Since: 1.2.0

runStateGen_ :: RandomGen g => g -> (StateGenM g -> State g a) -> a

Runs a monadic generating action in the State monad using a pure pseudo-random number generator. Returns only the resulting pseudo-random value.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> runStateGen_ pureGen randomM :: Int
7879794327570578227

Since: 1.2.0

runStateGenT :: RandomGen g => g -> (StateGenM g -> StateT g m a) -> m (a, g)

Runs a monadic generating action in the StateT monad using a pure pseudo-random number generator.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> runStateGenT pureGen randomM :: IO (Int, StdGen)
(7879794327570578227,StdGen {unStdGen = SMGen 11285859549637045894 7641485672361121627})

Since: 1.2.0

runStateGenT_ :: (RandomGen g, Functor f) => g -> (StateGenM g -> StateT g f a) -> f a

Runs a monadic generating action in the StateT monad using a pure pseudo-random number generator. Returns only the resulting pseudo-random value.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> runStateGenT_ pureGen randomM :: IO Int
7879794327570578227

Since: 1.2.1

runStateGenST :: RandomGen g => g -> (forall s. StateGenM g -> StateT g (ST s) a) -> (a, g)

Runs a monadic generating action in the ST monad using a pure pseudo-random number generator.

Since: 1.2.0

runStateGenST_ :: RandomGen g => g -> (forall s. StateGenM g -> StateT g (ST s) a) -> a

Runs a monadic generating action in the ST monad using a pure pseudo-random number generator. Same as runStateGenST, but discards the resulting generator.

Since: 1.2.1

Mutable adapter with atomic operations

newtype AtomicGen g

Frozen version of mutable AtomicGenM generator

Since: 1.2.0

Constructors

AtomicGen

Fields

unAtomicGen :: g

newtype AtomicGenM g

Wraps an IORef that holds a pure pseudo-random number generator. All operations are performed atomically.

AtomicGenM is safe in the presence of exceptions and concurrency.

AtomicGenM is the slowest of the monadic adapters due to the overhead of its atomic operations.

Since: 1.2.0

Constructors

AtomicGenM

Fields

unAtomicGenM :: IORef g

newAtomicGenM :: MonadIO m => g -> m (AtomicGenM g)

Creates a new AtomicGenM.

Since: 1.2.0

applyAtomicGen :: MonadIO m => (g -> (a, g)) -> AtomicGenM g -> m a

Atomically applies a pure operation to the wrapped pseudo-random number generator.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> g <- newAtomicGenM pureGen
>>> applyAtomicGen random g :: IO Int
7879794327570578227

Since: 1.2.0

globalStdGen :: AtomicGenM StdGen

Global mutable standard pseudo-random number generator. This is the same generator that was historically used by randomIO and randomRIO functions.

>>> replicateM 10 (uniformRM ('a', 'z') globalStdGen)
"tdzxhyfvgr"

Since: 1.2.1

Mutable adapter in IO

newtype IOGen g

Frozen version of mutable IOGenM generator

Since: 1.2.0

Constructors

IOGen

Fields

unIOGen :: g

newtype IOGenM g

Wraps an IORef that holds a pure pseudo-random number generator.

IOGenM is safe in the presence of exceptions, but not concurrency.

IOGenM is slower than StateGenM due to the extra pointer indirection.

IOGenM is faster than AtomicGenM since the IORef operations used by IOGenM are not atomic.

An example use case is writing pseudo-random bytes into a file:

>>> import UnliftIO.Temporary (withSystemTempFile)
>>> import Data.ByteString (hPutStr)
>>> let ioGen g = withSystemTempFile "foo.bin" $ \_ h -> uniformRM (0, 100) g >>= flip uniformByteStringM g >>= hPutStr h

and then run it:

>>> newIOGenM (mkStdGen 1729) >>= ioGen

Since: 1.2.0

Constructors

IOGenM

Fields

unIOGenM :: IORef g

newIOGenM :: MonadIO m => g -> m (IOGenM g)

Creates a new IOGenM.

Since: 1.2.0

applyIOGen :: MonadIO m => (g -> (a, g)) -> IOGenM g -> m a

Applies a pure operation to the wrapped pseudo-random number generator.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> g <- newIOGenM pureGen
>>> applyIOGen random g :: IO Int
7879794327570578227

Since: 1.2.0

Mutable adapter in ST

newtype STGen g

Frozen version of mutable STGenM generator

Since: 1.2.0

Constructors

STGen

Fields

unSTGen :: g

newtype STGenM g s

Wraps an STRef that holds a pure pseudo-random number generator.

STGenM is safe in the presence of exceptions, but not concurrency.

STGenM is slower than StateGenM due to the extra pointer indirection.

Since: 1.2.0

Constructors

STGenM

Fields

unSTGenM :: STRef s g

newSTGenM :: g -> ST s (STGenM g s)

Creates a new STGenM.

Since: 1.2.0

applySTGen :: (g -> (a, g)) -> STGenM g s -> ST s a

Applies a pure operation to the wrapped pseudo-random number generator.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> (runSTGen pureGen (\g -> applySTGen random g)) :: (Int, StdGen)
(7879794327570578227,StdGen {unStdGen = SMGen 11285859549637045894 7641485672361121627})

Since: 1.2.0

runSTGen :: RandomGen g => g -> (forall s. STGenM g s -> ST s a) -> (a, g)

Runs a monadic generating action in the ST monad using a pure pseudo-random number generator.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> (runSTGen pureGen (\g -> applySTGen random g)) :: (Int, StdGen)
(7879794327570578227,StdGen {unStdGen = SMGen 11285859549637045894 7641485672361121627})

Since: 1.2.0

runSTGen_ :: RandomGen g => g -> (forall s. STGenM g s -> ST s a) -> a

Runs a monadic generating action in the ST monad using a pure pseudo-random number generator. Returns only the resulting pseudo-random value.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> (runSTGen_ pureGen (\g -> applySTGen random g)) :: Int
7879794327570578227

Since: 1.2.0

Mutable adapter in STM

newtype TGen g

Frozen version of mutable TGenM generator

Since: 1.2.1

Constructors

TGen

Fields

unTGen :: g

newtype TGenM g

Wraps a TVar that holds a pure pseudo-random number generator.

Since: 1.2.1

Constructors

TGenM

Fields

unTGenM :: TVar g

newTGenM :: g -> STM (TGenM g)

Creates a new TGenM in STM.

Since: 1.2.1

newTGenMIO :: MonadIO m => g -> m (TGenM g)

Creates a new TGenM in IO.

Since: 1.2.1

applyTGen :: (g -> (a, g)) -> TGenM g -> STM a

Applies a pure operation to the wrapped pseudo-random number generator.

Examples

Expand

>>> import Control.Concurrent.STM
>>> import System.Random.Stateful
>>> import Data.Int (Int32)
>>> let pureGen = mkStdGen 137
>>> stmGen <- newTGenMIO pureGen
>>> atomically $ applyTGen uniform stmGen :: IO Int32
637238067

Since: 1.2.1

Pseudo-random values of various types

This library provides two type classes to generate pseudo-random values:

UniformRange is used to generate a value of a type uniformly within a range.

Uniform is used to generate a value of a type uniformly over all possible values of that type.

Types may have instances for both or just one of UniformRange and Uniform. A few examples illustrate this:

Int, Word16 and Bool are instances of both UniformRange and Uniform.

Integer, Float and Double each have an instance for UniformRange but no Uniform instance.

A hypothetical type Radian representing angles by taking values in the range [0, 2π) has a trivial Uniform instance, but no UniformRange instance: the problem is that two given Radian values always span two ranges, one clockwise and one anti-clockwise.

It is trivial to construct a Uniform (a, b) instance given Uniform a and Uniform b (and this library provides this tuple instance).

On the other hand, there is no correct way to construct a UniformRange (a, b) instance based on just UniformRange a and UniformRange b.

class Uniform a where

The class of types for which a uniformly distributed value can be drawn from all possible values of the type.

Since: 1.2.0

Minimal complete definition

Nothing

Methods

uniformM :: StatefulGen g m => g -> m a

Generates a value uniformly distributed over all possible values of that type.

There is a default implementation via Generic:

>>> :set -XDeriveGeneric -XDeriveAnyClass
>>> import GHC.Generics (Generic)
>>> import System.Random.Stateful
>>> data MyBool = MyTrue | MyFalse deriving (Show, Generic, Finite, Uniform)
>>> data Action = Code MyBool | Eat (Maybe Bool) | Sleep deriving (Show, Generic, Finite, Uniform)
>>> gen <- newIOGenM (mkStdGen 42)
>>> uniformListM 10 gen :: IO [Action]
[Code MyTrue,Code MyTrue,Eat Nothing,Code MyFalse,Eat (Just False),Eat (Just True),Eat Nothing,Eat (Just False),Sleep,Code MyFalse]

Since: 1.2.0

default uniformM :: (StatefulGen g m, Generic a, GUniform (Rep a)) => g -> m a

uniformListM :: (StatefulGen g m, Uniform a) => Int -> g -> m [a]

Generates a list of pseudo-random values.

Examples

Expand

>>> import System.Random.Stateful
>>> let pureGen = mkStdGen 137
>>> g <- newIOGenM pureGen
>>> uniformListM 10 g :: IO [Bool]
[True,True,True,True,False,True,True,False,False,False]

Since: 1.2.0

uniformViaFiniteM :: (StatefulGen g m, Generic a, GFinite (Rep a)) => g -> m a

A definition of Uniform for Finite types. If your data has several fields of sub-Word cardinality, this instance may be more efficient than one, derived via Generic and GUniform.

>>> :set -XDeriveGeneric -XDeriveAnyClass
>>> import GHC.Generics (Generic)
>>> import System.Random.Stateful
>>> data Triple = Triple Word8 Word8 Word8 deriving (Show, Generic, Finite)
>>> instance Uniform Triple where uniformM = uniformViaFiniteM
>>> gen <- newIOGenM (mkStdGen 42)
>>> uniformListM 5 gen :: IO [Triple]
[Triple 60 226 48,Triple 234 194 151,Triple 112 96 95,Triple 51 251 15,Triple 6 0 208]

class UniformRange a where

The class of types for which a uniformly distributed value can be drawn from a range.

Since: 1.2.0

Methods

uniformRM :: StatefulGen g m => (a, a) -> g -> m a

Generates a value uniformly distributed over the provided range, which is interpreted as inclusive in the lower and upper bound.

uniformRM (1 :: Int, 4 :: Int) generates values uniformly from the set \(\{1,2,3,4\}\)

uniformRM (1 :: Float, 4 :: Float) generates values uniformly from the set \(\{x\;|\;1 \le x \le 4\}\)

The following law should hold to make the function always defined:

uniformRM (a, b) = uniformRM (b, a)

Since: 1.2.0

Generators for sequences of pseudo-random bytes

genShortByteStringIO

Arguments

:: MonadIO m

=> Int
Number of bytes to generate

-> m Word64
IO action that can generate 8 random bytes at a time

-> m ShortByteString

Efficiently generates a sequence of pseudo-random bytes in a platform independent manner.

Since: 1.2.0

genShortByteStringST :: Int -> ST s Word64 -> ST s ShortByteString

Same as genShortByteStringIO, but runs in ST.

Since: 1.2.0

uniformByteStringM :: StatefulGen g m => Int -> g -> m ByteString

Generates a pseudo-random ByteString of the specified size.

Since: 1.2.0

uniformDouble01M :: forall g m. StatefulGen g m => g -> m Double

Generates uniformly distributed Double in the range \([0, 1]\). Numbers are generated by generating uniform Word64 and dividing it by \(2^{64}\). It's used to implement UniformRange instance for Double.

Since: 1.2.0

uniformDoublePositive01M :: forall g m. StatefulGen g m => g -> m Double

Generates uniformly distributed Double in the range \((0, 1]\). Number is generated as \(2^{-64}/2+\operatorname{uniformDouble01M}\). Constant is 1/2 of smallest nonzero value which could be generated by uniformDouble01M.

Since: 1.2.0

uniformFloat01M :: forall g m. StatefulGen g m => g -> m Float

Generates uniformly distributed Float in the range \([0, 1]\). Numbers are generated by generating uniform Word32 and dividing it by \(2^{32}\). It's used to implement UniformRange instance for Float.

Since: 1.2.0

uniformFloatPositive01M :: forall g m. StatefulGen g m => g -> m Float

Generates uniformly distributed Float in the range \((0, 1]\). Number is generated as \(2^{-32}/2+\operatorname{uniformFloat01M}\). Constant is 1/2 of smallest nonzero value which could be generated by uniformFloat01M.

Since: 1.2.0

uniformEnumM :: forall a g m. (Enum a, Bounded a, StatefulGen g m) => g -> m a

Generates uniformly distributed Enum. One can use it to define a Uniform instance:

data Colors = Red | Green | Blue deriving (Enum, Bounded)
instance Uniform Colors where uniformM = uniformEnumM

Since: 1.2.1

uniformEnumRM :: forall a g m. (Enum a, StatefulGen g m) => (a, a) -> g -> m a

Generates uniformly distributed Enum in the given range. One can use it to define a UniformRange instance:

data Colors = Red | Green | Blue deriving (Enum)
instance UniformRange Colors where
  uniformRM = uniformEnumRM
  inInRange (lo, hi) x = isInRange (fromEnum lo, fromEnum hi) (fromEnum x)

Since: 1.2.1

Appendix

How to implement StatefulGen

Typically, a monadic pseudo-random number generator has facilities to save and restore its internal state in addition to generating pseudo-random numbers.

Here is an example instance for the monadic pseudo-random number generator from the mwc-random package:

instance (s ~ PrimState m, PrimMonad m) => StatefulGen (MWC.Gen s) m where
  uniformWord8 = MWC.uniform
  uniformWord16 = MWC.uniform
  uniformWord32 = MWC.uniform
  uniformWord64 = MWC.uniform
  uniformShortByteString n g = unsafeSTToPrim (genShortByteStringST n (MWC.uniform g))

instance PrimMonad m => FrozenGen MWC.Seed m where
  type MutableGen MWC.Seed m = MWC.Gen (PrimState m)
  thawGen = MWC.restore
  freezeGen = MWC.save

FrozenGen

FrozenGen gives us ability to use any stateful pseudo-random number generator in its immutable form, if one exists that is. This concept is commonly known as a seed, which allows us to save and restore the actual mutable state of a pseudo-random number generator. The biggest benefit that can be drawn from a polymorphic access to a stateful pseudo-random number generator in a frozen form is the ability to serialize, deserialize and possibly even use the stateful generator in a pure setting without knowing the actual type of a generator ahead of time. For example we can write a function that accepts a frozen state of some pseudo-random number generator and produces a short list with random even integers.

>>> import Data.Int (Int8)
>>> :{
myCustomRandomList :: FrozenGen f m => f -> m [Int8]
myCustomRandomList f =
  withMutableGen_ f $ \gen -> do
    len <- uniformRM (5, 10) gen
    replicateM len $ do
      x <- uniformM gen
      pure $ if even x then x else x + 1
:}

and later we can apply it to a frozen version of a stateful generator, such as STGen:

>>> print $ runST $ myCustomRandomList (STGen (mkStdGen 217))
[-50,-2,4,-8,-58,-40,24,-32,-110,24]

or a Seed from mwc-random:

>>> import Data.Vector.Primitive as P
>>> print $ runST $ myCustomRandomList (MWC.toSeed (P.fromList [1,2,3]))
[24,40,10,40,-8,48,-78,70,-12]

Alternatively, instead of discarding the final state of the generator, as it happens above, we could have used withMutableGen, which together with the result would give us back its frozen form. This would allow us to store the end state of our generator somewhere for the later reuse.

Floating point number caveats

The UniformRange instances for Float and Double use the following procedure to generate a random value in a range for uniformRM (a, b) g:

If \(a = b\), return \(a\). Otherwise:

Generate \(x\) uniformly such that \(0 \leq x \leq 1\).

The method by which \(x\) is sampled does not cover all representable floating point numbers in the unit interval. The method never generates denormal floating point numbers, for example.

Return \(x \cdot a + (1 - x) \cdot b\).

Due to rounding errors, floating point operations are neither associative nor distributive the way the corresponding operations on real numbers are. Additionally, floating point numbers admit special values NaN as well as negative and positive infinity.

For pathological values, step 2 can yield surprising results.

The result may be greater than max a b.

>>> :{
let (a, b, x) = (-2.13238e-29, -2.1323799e-29, 0.27736077)
    result = x * a + (1 - x) * b :: Float
in (result, result > max a b)
:}
(-2.1323797e-29,True)

The result may be smaller than min a b.

>>> :{
let (a, b, x) = (-1.9087862, -1.908786, 0.4228573)
    result = x * a + (1 - x) * b :: Float
in (result, result < min a b)
:}
(-1.9087863,True)

What happens when NaN or Infinity are given to uniformRM? We first define them as constants:

>>> nan = read "NaN" :: Float
>>> inf = read "Infinity" :: Float

If at least one of \(a\) or \(b\) is NaN, the result is NaN.

>>> let (a, b, x) = (nan, 1, 0.5) in x * a + (1 - x) * b
NaN
>>> let (a, b, x) = (-1, nan, 0.5) in x * a + (1 - x) * b
NaN

If \(a\) is -Infinity and \(b\) is Infinity, the result is NaN.

>>> let (a, b, x) = (-inf, inf, 0.5) in x * a + (1 - x) * b
NaN

Otherwise, if \(a\) is Infinity or -Infinity, the result is \(a\).

>>> let (a, b, x) = (inf, 1, 0.5) in x * a + (1 - x) * b
Infinity
>>> let (a, b, x) = (-inf, 1, 0.5) in x * a + (1 - x) * b
-Infinity

Otherwise, if \(b\) is Infinity or -Infinity, the result is \(b\).

>>> let (a, b, x) = (1, inf, 0.5) in x * a + (1 - x) * b
Infinity
>>> let (a, b, x) = (1, -inf, 0.5) in x * a + (1 - x) * b
-Infinity

Note that the GCC 10.1.0 C++ standard library, the Java 10 standard library and CPython 3.8 use the same procedure to generate floating point values in a range.

References

Guy L. Steele, Jr., Doug Lea, and Christine H. Flood. 2014. Fast splittable pseudorandom number generators. In Proceedings of the 2014 ACM International Conference on Object Oriented Programming Systems Languages & Applications (OOPSLA '14). ACM, New York, NY, USA, 453-472. DOI: https://doi.org/10.1145/2660193.2660195